🚚 Move: userProductService 삭제
